Create 10_supermarket_receipt_generator
This code will generate a receipt for input groceries item, list out items that are out of stock, and finally printing sold out items also. This code can be refactored and can include a dictionary that maps items with code as well. Refactoring of some loops can also be done.
Hopefully, I'll do the required updates. Also, the usage of functions, classes are also pending to improve the code. It'll reflect in the upcoming changes.