Create 10_replace_set_of_strings_with_another
This code will take user input of string and then replaces the first occurrence of a set of words 'not' followed by 'poor' with the word 'good'. This code can be improved by changing every set of the occurrence that we don't want. With the help of a loop, that can be achieved.
Hopefully, I'll do the required updates. It'll reflect in the upcoming changes.