Create 11_check_leap_year
This code checks whether the input year is a leap year or not. This code can be refactored to handle errors viz if user input alphabets, or special characters.
Hopefully, I'll do the required updates. It'll reflect in the upcoming changes along with further use of this piece of code.