Create 13_reading_sem_marks.csv
This code read and extract information present in the CSV file. It prints out the total number of rows, field names and finally a list of students followed by their semester and marks. This piece of code can be loaded with functions which provide different operations viz highest scorer in all semesters, highest scorer overall, students failing, the number of students in each semester, etc.
Hopefully, in the upcoming commits, I may add functions and a menu for the user to select the operation.